// src/proto.d.ts
export namespace routerrpc {
// ...
    interface ICircuitKey {

        /** The id of the channel that the is part of this circuit. */
        chanId?: (Long|null);

        /** The index of the incoming htlc in the incoming channel. */
        htlcId?: (Long|null);
    }
// ...
    interface IForwardHtlcInterceptRequest {

        /** ForwardHtlcInterceptRequest incomingCircuitKey */
        incomingCircuitKey?: (routerrpc.ICircuitKey|null);
// ...
    }
// ...
}